# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FinFlow API",
    "version": "1.0.0",
    "description": "All endpoints require an authenticated session (Flask-Login cookie, set by /login) and are scoped to the logged-in user's own accounts. State-changing requests (POST/PATCH/DELETE) also require a CSRF token, sent as header `X-CSRFToken` (get one from the `csrf-token` <meta> tag on any page, or the `csrf_token` hidden field on the login/signup forms)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "parameters": {
      "CsrfHeader": {
        "name": "X-CSRFToken",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Required on POST/PATCH/DELETE."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/what-if/budget": {
      "post": {
        "summary": "Simulate reducing spend in specific categories",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "add_savings": {
                    "type": "number",
                    "default": 0
                  },
                  "reduce_categories": {
                    "type": "object",
                    "description": "category_id -> reduction fraction (0-1). Keys are whitelisted against dim_category.",
                    "additionalProperties": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "monthly_savings_impact": {
                      "type": "number"
                    },
                    "projected_balance_6_months": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown category key(s)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/what-if/investments": {
      "post": {
        "summary": "Project a lump-sum investment at a blended return rate",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfHeader"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "invest_amount": {
                    "type": "number"
                  },
                  "allocation": {
                    "type": "object",
                    "properties": {
                      "stocks": {
                        "type": "number"
                      },
                      "bonds": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projected_1_year": {
                      "type": "number"
                    },
                    "projected_5_years": {
                      "type": "number"
                    },
                    "projected_10_years": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}